Balance changes + Energy fixes
Player stats and bullet stats have been changed
Energy now depleets properly

+++ Year3CapstoneProject/Assets/Scripts/UI/Bars.cs
@@ -22,7 +22,8 @@ public class Bars : MonoBehaviour
 
     private void UseEnergy()
     {
-        energyBar.fillAmount = playerStats.CurrentEnergy/playerStats.MaxEnergy;
+
+        energyBar.fillAmount = (float) playerStats.CurrentEnergy /(float)playerStats.MaxEnergy;
     }
 
     private void Update()
